requirements.txt
anyio==4.1.0
APScheduler==3.10.4
certifi==2023.11.17
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
tzdata==2023.3
tzlocal==5.2
